+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # CHANGELOG
 
+## 0.1.1
+
+- Bugfix: remove the "is-plain-object" dependency and replace it with a simple object check as class instances should be treated as objects as well but are not "plain" objects
+- Add example for a NestJS webserver using the ts-configurable package
+- Update documentation
+
 ## 0.1.0
 
 - Update documentation